Uses static_cast when converting from float to double

+++ Source/AGXUnreal/Private/Constraints/Controllers/AGX_TargetSpeedController.cpp
@@ -26,8 +26,8 @@ void FAGX_ConstraintTargetSpeedController::ToBarrier(FTargetSpeedControllerBarri
 	Barrier->bEnable = bEnable;
 	Barrier->Elasticity = Elasticity;
 	Barrier->Damping = Damping;
-	Barrier->ForceRangeMin = ForceRange.Min;
-	Barrier->ForceRangeMax = ForceRange.Max;
+	Barrier->ForceRangeMin = static_cast<double>(ForceRange.Min);
+	Barrier->ForceRangeMax = static_cast<double>(ForceRange.Max);
 
 	Barrier->bRotational = bRotational;
